+++ NEWS.md
@@ -1,3 +1,11 @@
+
+# carelesswhisper 0.1.1  2023-06-17
+
+* Added `details` option to `whisper()` to have it return a data.frame of details on
+  tokens, rather than just the final text string.
+    * Internal copy of [`{deft}`](https://github.com/coolbutuseless/deft) package for data.frame creation in C
+* Added `jfk` audio sample for testing (copied from `whisper.cpp`)
+
 # carelesswhisper 0.1.0
 
 * Initial release